Refactor memory service and remove ports
