+++ ciris_engine/memory/__init__.py
@@ -2,8 +2,16 @@
 
 from .utils import is_wa_feedback
 from .memory_handler import classify_target
+from ciris_engine.adapters.local_graph_memory import (
+    LocalGraphMemoryService,
+    MemoryOpResult,
+    MemoryOpStatus,
+)
 
 __all__ = [
     "is_wa_feedback",
     "classify_target",
+    "LocalGraphMemoryService",
+    "MemoryOpResult",
+    "MemoryOpStatus",
 ]
