--- ciris_engine/ports/__init__.py
@@ -1,6 +1,0 @@
-from .ports import ActionSink, DeferralSink, EventSource, FeedbackSink
-from .service_aware_sinks import (
-    ServiceAwareActionSink, 
-    ServiceAwareDeferralSink, 
-    ExampleDiscordActionSink
-)
