+++ tests/test_memory_integration.py
@@ -1,11 +1,11 @@
 import asyncio
 import pytest
-from ciris_engine.memory.ciris_local_graph import CIRISLocalGraph
+from ciris_engine.adapters.local_graph_memory import LocalGraphMemoryService
 from ciris_engine.schemas.graph_schemas_v1 import GraphNode, NodeType, GraphScope
 
 @pytest.mark.asyncio
 async def test_memory_operations():
-    memory = CIRISLocalGraph(":memory:")
+    memory = LocalGraphMemoryService(":memory:")
     await memory.start()
     
     # Test memorize
